pyproject.toml
[tool.poetry]
name = "libertai-inference"
version = "0.0.1"
description = ""
authors = ["Reza Rahemtola <49811529+RezaRahemtola@users.noreply.github.com>"]
readme = "README.md"
homepage = "https://libertai.io"
repository = "https://github.com/LibertAI/libertai-subscriptions"
documentation = "https://docs.libertai.io"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = { extras = ["standard"], version = "^0.114.1" }
apscheduler = "^3.10.4"
setuptools = "^75.4.0"
web3 = "^7.9.0"
requests = "^2.32.3"
sqlalchemy = { extras = ["mypy"], version = "^2.0.39" }
alembic = "^1.15.1"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
pyjwt = "^2.8.0"
libertai-utils = "^0.0.13"

[tool.poetry.group.dev.dependencies]
mypy = "^1.11.1"
ruff = "^0.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"